refactor: DIP 원칙에 맞도록 vendingMachine 클래스가 추상화 모듈을 의존하도록 변경
