+++ src/model/product.js
@@ -1,7 +1,7 @@
 export class Product {
-    constructor(name, quantity, price) {
-        this.name = name;
-        this.price = price;
-        this.quantity = quantity;
-    }
+	constructor(name, quantity, price) {
+		this.name = name;
+		this.price = price;
+		this.quantity = quantity;
+	}
 }
